Rebuilt random to use a beta distribution
random_int_bell_curve now generates based on a beta distribution. Alpha
and Beta are tuned by lean and narrow, which control the mean of the
distribution and breadth of the bell curve. Use print_histogram to
experiment with values.
print_histogram is now significantly more efficient, and accepts an
iterator, as well as optional left and right bounds for the graph